kvstorage: mv TestIterateIDPrefixKeys to kvstorage
The function that this test exercised lives in the kvstorage package.

Epic: none
Release note: none
