+++ pkg/kv/kvserver/kvstorage/BUILD.bazel
@@ -42,6 +42,7 @@ go_test(
         "cluster_version_test.go",
         "datadriven_test.go",
         "destroy_test.go",
+        "init_test.go",
         "initial_test.go",
         "stateloader_test.go",
     ],
@@ -66,11 +67,13 @@ go_test(
         "//pkg/util/hlc",
         "//pkg/util/leaktest",
         "//pkg/util/log",
+        "//pkg/util/randutil",
         "//pkg/util/stop",
         "//pkg/util/tracing",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_datadriven//:datadriven",
         "@com_github_cockroachdb_redact//:redact",
+        "@com_github_kr_pretty//:pretty",
         "@com_github_stretchr_testify//require",
     ],
 )
